Aggregated sampling of stablesr

+++ train_images_vae_fp16.py
@@ -50,7 +50,7 @@ def update_ema(ema_model, model, decay=0.999):
         ema_params[name].mul_(decay).add_(param.data, alpha=1 - decay)
 
 t32 = lambda x: x.to(torch.float32)
-t16 = lambda x: x.to(torch.float16)
+t16 = lambda x: x.to(torch.bfloat16)
 
 def requires_grad(model, flag=True):
     """
@@ -158,7 +158,7 @@ def main(args):
     requires_grad(ema, False)
     model = DDP(model.to(device), device_ids=[rank])
     diffusion = create_diffusion(timestep_respacing="", diffusion_steps=1000)  # default: 1000 steps, linear noise schedule
-    vae = AutoencoderKL.from_pretrained(f"./vae", torch_dtype=torch.float16).to(device)
+    vae = AutoencoderKL.from_pretrained(f"./vae", torch_dtype=torch.bfloat16).to(device)
     logger.info(f"DiT Parameters: {sum(p.numel() for p in model.parameters()):,}")
 
     # Setup optimizer (we used default Adam betas=(0.9, 0.999) and a constant learning rate of 1e-4 in our paper):
